update
Former-commit-id: 31a1bdf02e578ff63ac5ed8357cc1823df9389d7

+++ src/server/table/TableQuery.java
@@ -224,7 +224,7 @@ public class TableQuery {
 
         // linking threshold
         double linkingThreshold = additionalParameters != null && additionalParameters.containsKey("linking-threshold")
-                ? (float) additionalParameters.get("linking-threshold") : -1; // default is no-threshold
+                ? (double) additionalParameters.get("linking-threshold") : -1; // default is no-threshold
 
         // retrieve additional parameters
         Session session = additionalParameters == null ? null : (Session) additionalParameters.get("session");
